+++ plugins/dialog/src/error.rs
@@ -18,9 +18,6 @@ pub enum Error {
     #[cfg(mobile)]
     #[error("Folder picker is not implemented on mobile")]
     FolderPickerNotImplemented,
-    #[cfg(target_os = "ios")]
-    #[error("File save dialog is not implemented on iOS")]
-    FileSaveDialogNotImplemented,
     #[error(transparent)]
     Fs(#[from] tauri_plugin_fs::Error),
     #[error("URL is not a valid path")]